requirements.txt
gitpython>=3.1.30


matplotlib>=3.3​


numpy>=1.23.5​


opencv-python>=4.1.1​


pillow>=10.3.0​


psutil​


PyYAML>=5.3.1​

DagsHub


requests>=2.32.2​


scipy>=1.4.1​

DagsHub


thop>=0.1.1​


torch>=1.8.0​


torchvision>=0.9.0​


tqdm>=4.66.3​


ultralytics>=8.2.34